growing-food-api/requirements.txt
Django==3.1.2
psycopg2-binary==2.8.6
djangorestframework==3.13.1
black==22.3.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2